Remove compilation warnings from db load/save

+++ db.h
@@ -306,7 +306,7 @@ class	object
 			void	set_flag			(const object_flag_type f)	{ flags[f/8] |= (flag_type)(1<<(f&7)); }
 			void	clear_flag			(const object_flag_type f)	{ flags[f/8] &= (flag_type)(~(1<<(f&7))); }
 			bool	get_flag			(const object_flag_type f)	{ return (flags[f/8] & (flag_type)(1<<(f&7))) != 0; }
-			void	set_flag_byte			(const int c, const flag_type v){ flags[c] = v; }
+			void	set_flag_byte			(const size_t c, const flag_type v){ flags[c] = v; }
 			flag_type	get_flag_byte		(const int c)			const { return (flags[c]); }
 
 			void	set_referenced			()				{ set_flag(FLAG_REFERENCED); }
